Code design changed
