+++ src/common/fab/stylables.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components/native';
+
+export const FabStyle = styled.TouchableOpacity.attrs(props => {})`
+  height: 48px;
+  padding-left: 16px;
+  padding-right: 16px;
+  justify-content: center;
+  align-items: center;
+  border-radius: 32px;
+  background-color: green;
+`;
+
+export const FabText = styled.Text`
+  color: white;
+  font-size: 16px;
+`;
