+++ src/common/progress/ButtonProgress.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import * as Progress from 'react-native-progress';
+import { ProgressBox } from './stylables';
+
+export default ButtonProgress = ({color}) => (
+  <ProgressBox color={color}>
+    <Progress.Circle size={30} indeterminate={true} color="white" />
+  </ProgressBox>ø
+);
