+++ src/common/roundedbutton/RoundedButton.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import {RoundedTouchableOpacity, TextButton} from './stylables';
+
+export const RoundedButton = ({onPress, title, color}) => (
+  <RoundedTouchableOpacity onPress={onPress} color={color} activeOpacity={0.5}>
+    <TextButton>{title}</TextButton>
+  </RoundedTouchableOpacity>
+);
