+++ src/common/screentitle/ScreenTitle.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import {TitleTextStyle} from './stylables';
+
+export const ScreenTitle = props => (
+  <TitleTextStyle>{props.title}</TitleTextStyle>
+);
