+++ src/common/screentitle/stylables.js
@@ -1,8 +1,7 @@
-import React from 'react';
 import {Text} from 'react-native';
 import styled from 'styled-components';
 
-const TitleTextStyle = styled.Text`
+export const TitleTextStyle = styled.Text`
   color: black;
   font-size: 32px;
   font-weight: 500;
@@ -10,7 +9,3 @@ const TitleTextStyle = styled.Text`
   margin-left: 16px;
   margin-bottom: 16px;
 `;
-
-export const ScreenTitle = props => (
-  <TitleTextStyle>{props.title}</TitleTextStyle>
-);
