+++ src/ui/auth/login/index.js
@@ -1,0 +1,1 @@
+import Login from './Login';
