+++ src/ui/home/dashboard/index.js
@@ -1,0 +1,1 @@
+import Home from './Home';
